Finished power up system implementation. Currently there is still one power up that is not fully implemented yet. Also updated bullet speed and bounding box size. The power up spawning system still needs to be implimented

+++ Nian-Gao/Assets/Scripts/SwitcherManager.cs
@@ -9,7 +9,7 @@ public class SwitcherManager : MonoBehaviour
     //keeps track of the last level played
     public int lastLevel = 0;
     //keeps track of player health
-    public int playerHealth = 0;
+    public float playerHealth = 0;
 
     //add references to anything else we want between levels here
